+++ src/components/FlawFormOwner.vue
@@ -1,7 +1,13 @@
 <script setup lang="ts">
-import EditableText from '@/components/widgets/EditableText.vue';
+import { nextTick, computed, ref, toValue } from 'vue';
+import sanitize from 'sanitize-html';
+import { useMemoize } from '@vueuse/core';
+import { createCatchHandler } from '@/composables/service-helpers';
+import { searchJiraUsers } from '@/services/JiraService';
 import { useUserStore } from '@/stores/UserStore';
-import { nextTick, computed, ref } from 'vue';
+import type { ZodJiraUserPickerType } from '@/types/zodJira';
+import LabelDiv from './widgets/LabelDiv.vue';
+import EditableTextWithSuggestions from './widgets/EditableTextWithSuggestions.vue';
 
 const owner = defineModel<string | null>({ default: null });
 const userStore = useUserStore();
@@ -17,7 +23,7 @@ async function selfAssign() {
   });
 }
 
-async function handleClick(fn: (arg?: any) => any){
+async function handleClick(fn: (arg?: any) => any) {
   await selfAssign();
   nextTick(fn);
 }
@@ -27,44 +33,65 @@ const isAssignedToMe = computed(() =>
 );
 
 const isLoading = ref(false);
+const results = ref<ZodJiraUserPickerType[]>([]);
 
+const cacheJiraUsers = useMemoize(async (query: string) => {
+  if (!query) {
+    return [];
+  }
+  isLoading.value = true;
+  const users = await searchJiraUsers(query)
+    .catch(createCatchHandler('Failed to search jira users', false));
+  isLoading.value = false;
+  return users?.data?.users ?? [];
+});
+
+
+const onQueryChange = async (query: string) => {
+  const users = await cacheJiraUsers(query);
+  results.value = users ?? [];
+};
+
+const handleSuggestionClick = (fn: (args?: any) => void, user: ZodJiraUserPickerType) => {
+  owner.value = toValue(user.name);
+  results.value = [];
+  nextTick(fn);
+};
 </script>
 
 <template>
-  <label class="ps-3 mb-2 input-group osim-input">
-    <div class="row">
-      <span class="form-label col-3 pe-3 ">
-        Owner
-      </span>
-      <EditableText v-model="owner">
-        <template v-if="!isAssignedToMe" #buttons-out-of-editing-mode="{ onBlur }">
-          <button
-            type="button"
-            class="btn btn-primary osim-self-assign"
-            :disabled="isLoading"
-            @click.prevent.stop="handleClick(onBlur)"
-          >
-            Self Assign
-          </button>
-        </template>
-        <template v-if="!isAssignedToMe" #buttons-in-editing-mode="{ onBlur }">
-          <button
-            type="button"
-            class="btn btn-primary osim-self-assign"
-            :disabled="isLoading"
-            @click.prevent.stop="handleClick(onBlur)"
-          >
-            Self Assign
-          </button>
-        </template>
-      </EditableText>
-    </div>
-  </label>
+  <LabelDiv label="Owner" :loading="isLoading">
+    <EditableTextWithSuggestions v-model="owner" class="col-12" @update:query="onQueryChange">
+      <template v-if="!isAssignedToMe" #buttons-out-of-editing-mode="{ onBlur }">
+        <button
+          type="button"
+          class="btn btn-primary osim-self-assign"
+          :disabled="isLoading"
+          @click.prevent.stop="handleClick(onBlur)"
+        >
+          Self Assign
+        </button>
+      </template>
+      <template v-if="!isAssignedToMe" #buttons-in-editing-mode="{ onBlur }">
+        <button
+          type="button"
+          class="btn btn-primary osim-self-assign"
+          :disabled="isLoading"
+          @click.prevent.stop="handleClick(onBlur)"
+        >
+          Self Assign
+        </button>
+      </template>
+      <template v-if="results.length > 0" #suggestions="{ onBlur }">
+        <div
+          v-for="user in results"
+          :key="user.name"
+          class="item"
+          @click.prevent.stop="handleSuggestionClick(onBlur, user)"
+        >
+          <span v-html="sanitize(user.html)" />
+        </div>
+      </template>
+    </EditableTextWithSuggestions>
+  </LabelDiv>
 </template>
-
-
-<style lang="scss" scoped>
-label.osim-input {
-  display: block;
-}
-</style>